create form and delete logic
Each form component excluding the PersonalForm now have a function to create a form. When a user wants to add an item to their resume, they will need to create a new form with a unique ID. This ID will also be assigned to the item they create when clicking submit on the form. The purpose of this is to let the user later edit their item in the resume.

Because the number of forms is dependent on the users input, they have their own useState